services/email_service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
redis>=4.5.5
google-api-python-client>=2.85.0
google-auth>=2.16.0
google-auth-oauthlib>=0.5.3
httpx>=0.24.0
python-dotenv>=1.0.0
aiohttp>=3.8.4
pika>=1.3.2
aio-pika>=9.3.0